First attempt at implementing video merging
Creates a Swift Package for this library.
Implemented a single public method that receives a list of input files and merges them together.
It is not yet possible to specify the output file path, quality, or any other parameters.
It also does not report processing progress or errors.

Tested on five video files generated from the same camera, which is working.
This test was done using a Mac test app with a simple button and hard-coded paths to input files.
The source code was in the class directly in the app project and was copied over to the Swift Package project after testing.